OzonApi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
python-dateutil==2.8.2
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
python-multipart==0.0.6
aiofiles==23.2.1
matplotlib==3.7.2
seaborn==0.12.2